+++ app/src/main/java/ua/com/radiokot/photoprism/features/envconnection/di/EnvConnectionFeatureModules.kt
@@ -116,7 +116,7 @@ val envConnectionFeatureModules: List<Module> = listOf(
                     }
                 },
                 clientCertificatesGuideUrl = getProperty("clientCertificatesGuideUrl"),
-                rootUrlGuideUrl = getProperty("rootUrlGuideUrl"),
+                rootUrlGuideUrl = getProperty("connectionGuideUrl"),
             )
         }
     },