+++ app/src/main/java/ua/com/radiokot/photoprism/features/envconnection/view/model/EnvConnectionViewModel.kt
@@ -143,7 +143,7 @@ class EnvConnectionViewModel(
 
     fun onRootUrlGuideButtonClicked() {
         eventsSubject.onNext(
-            Event.OpenRootUrlGuide(
+            Event.OpenConnectionGuide(
                 url = rootUrlGuideUrl,
             )
         )
@@ -243,7 +243,7 @@ class EnvConnectionViewModel(
         object GoToGallery : Event
         object ChooseClientCertificateAlias : Event
         object ShowMissingClientCertificatesNotice : Event
-        class OpenRootUrlGuide(val url: String) : Event
+        class OpenConnectionGuide(val url: String) : Event
         class OpenClientCertificateGuide(val url: String) : Event
     }
 }